Rename mod `etype` to `bare`.
