+++ parser/src/error.rs
@@ -1,5 +1,5 @@
-mod etype;
+mod bare;
 mod set;
 
-pub use self::etype::{BareError, Label, Span};
+pub use self::bare::{BareError, Label, Span};
 pub use self::set::{ErrorSet, Errors};
