+++ parser/src/error/bare.rs
